medBlog/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.3
django-ckeditor==5.9.0
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-froala-editor==3.1.0
django-js-asset==1.2.2
idna==2.8
itsdangerous==1.1.0
MarkupSafe==1.1.1
pytz==2019.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
Werkzeug==0.16.0
